Split off RssLog::Scopes
